Add wrong checkout output

+++ tools/migrate_kong.py
@@ -1,11 +1,8 @@
 #!/bin/env python
 # coding:utf8
 
-
-import os
 import re
 import time
-import json
 import pickle
 import argparse
 import requests
@@ -25,14 +22,14 @@ def change_upstream(upstream_url, new_profile):
         profile = matchs[0]
         return upstream_url.replace(profile, new_profile).encode('utf8').replace("{0}uct".format(new_profile), "product")
     elif 'app' in upstream_url:
-       router = upstream_url.split(".")[0]
-       router_reg = re.compile(r"http://(.*)-(\w+-\w+)")
-       app, project = router_reg.findall(router)[0]
-       project_profile = re.findall(r'\w+-(\w+)', project)[0]
-       new_project = project.replace(project_profile, new_profile) 
-       return 'http://{0}.{1}:8080'.format(app, new_project).replace("{0}uct".format(new_profile), "product")
+        router = upstream_url.split(".")[0]
+        router_reg = re.compile(r"http://(.*)-(\w+-\w+)")
+        app, project = router_reg.findall(router)[0]
+        project_profile = re.findall(r'\w+-(\w+)', project)[0]
+        new_project = project.replace(project_profile, new_profile)
+        return 'http://{0}.{1}:8080'.format(app, new_project).replace("{0}uct".format(new_profile), "product")
     else:
-       return upstream_url.encode('utf8').replace("{0}uct".format(new_profile), "product")
+        return upstream_url.encode('utf8').replace("{0}uct".format(new_profile), "product")
         
 
 class Kong(object):
@@ -40,44 +37,53 @@ class Kong(object):
         self.admin_url = admin_url
 
     def export(self, filename):
-        url = '{0}/apis?size=2000'.format(self.admin_url)
-        resp = requests.get(url)
-        if resp.status_code == 200:
-            with open(filename, 'wb') as f:
-                pickle.dump(resp.json(), f)
-            return True
-        else:
+        try:
+            url = '{0}/apis?size=2000'.format(self.admin_url)
+            resp = requests.get(url)
+            if resp.status_code == 200:
+                with open(filename, 'wb') as f:
+                    pickle.dump(resp.json(), f)
+                return True
+            else:
+                pprint("Oops, some wring in here ,The response code is {}".format(resp.status_code))
+        except Exception as e:
             pprint('Export Failed')
-    # @staticmethod
+            pprint(e)
+
     def cleanall(self):
         resp = requests.get('{0}/apis?size=1000'.format(self.admin_url)).json()
         apis = resp.get('data')
+        if not apis:
+            pprint("The Kong admin seem no data yet")
+            exit(1)
         for api in apis:
             name = api.get('name')
             resp = requests.delete('{0}/apis/{1}'.format(self.admin_url, name))
             if resp.status_code != 204:
-                print '{0} delete faild'.format(name)        
+                pprint('{0} delete failed'.format(name))
 
-        
     def _create(self, api_data):
         create_url = '{0}/apis/'.format(self.admin_url)
         resp = requests.post(create_url, api_data)
         if resp.status_code != 201:
-           pprint('{0}'.format(resp.text))
+            pprint('{0}'.format(resp.text))
 
     def dump(self, filename):
-        with open(filename, 'rb') as f:
-            resp = pickle.load(f)
-        for api in resp.get('data'):
-            hosts = list()
-            uris = list()
-            orig_hosts = api.get('hosts')
-            hosts.append(','.join(orig_hosts))
-            orig_uris = api.get('uris')
-            uris.append(','.join(orig_uris))
-            api['hosts'] = hosts
-            api['uris'] = uris
-            self._create(api)
+        try:
+            with open(filename, 'rb') as f:
+                resp = pickle.load(f)
+            for api in resp.get('data'):
+                hosts = list()
+                uris = list()
+                orig_hosts = api.get('hosts')
+                hosts.append(','.join(orig_hosts))
+                orig_uris = api.get('uris')
+                uris.append(','.join(orig_uris))
+                api['hosts'] = hosts
+                api['uris'] = uris
+                self._create(api)
+        except Exception as e:
+            pprint("Oops, some wrong occurd {}, May be the file not dump file".format(e))
 
     def migrate(self, filename, host, project, profile, prefix):
         hosts = list()
@@ -86,8 +92,11 @@ class Kong(object):
             resp = pickle.load(f)
         base_projects = [ 
             "moses", "cmbs", "fastdfs", "kong",
-             "unauthenticated", "hidevopsio"
+            "unauthenticated", "hidevopsio"
         ]
+        if not resp.get('data'):
+            pprint("The dump file semm not data yet, Please check it")
+            exit(2)
         for api in resp.get('data'):
             uris = list()
             old_uris = api.get('uris')
@@ -103,14 +112,14 @@ class Kong(object):
                 self._create(api)
             else:
                 base_projects.append(project)
-                if any([ prefix in url.encode('utf8') 
-                    for prefix in base_projects for url in uris ]):
-                        self._create(api)
-                        pprint(new_upstream_url)
+                if any([prefix in url.encode('utf8')
+                        for prefix in base_projects
+                        for url in uris]):
+                    self._create(api)
+                    pprint(new_upstream_url)
 
 
 class Check(object):
-    "Check input args"
     def __init__(self, arg_ns):
         self.arg_ns = arg_ns
 
@@ -119,35 +128,31 @@ class Check(object):
         src = self.arg_ns.src
         dest = self.arg_ns.dest
         host = self.arg_ns.host
-        project = self.arg_ns.project
-        profile = self.arg_ns.profile
         filename = self.arg_ns.filename
-        prefix = self.arg_ns.prefix
-        if filename:
+        if action == 'import' or action == 'export' or action == 'migrate':
+            if not filename:
+                pprint("--filename must exists!")
+                return False
             if action == 'export':
-                if src:
-                    return True
-                else:
-                    pprint('when action is export ,the --src is must args')
+                if not src:
+                    pprint("when action is export, --src must exists")
                     return False
-            elif action == 'migrate':
-                if dest and host:
-                    return True
-                else:
-                    pprint('when action is import, the --dest and --host is must args')
+            if action == 'migrate':
+                if not (dest and host):
+                    pprint("when action is migrate, --dest and --host must exists")
                     return False
-            elif action == 'import':
-                if dest:
-                    return True
-                else:
-                    pprint('when import ,the --dest is must args')
-            elif action == 'cleanall':
-                if dest:
-                    return True
+            if action == 'import':
+                if not dest:
+                    pprint("when action is import, --dest must exists")
+                    return False
+        elif action == 'cleanall':
+            if not dest:
+                pprint("when action is cleanall, --dest must exists")
+                return False
         else:
-            pprint("The args --filename must exist!")
-            pprint("use -h help")
+            pprint("not support action, use -h")
             return False
+        return True
 
 
 if __name__ == '__main__':
@@ -158,25 +163,27 @@ if __name__ == '__main__':
     kong_group.add_argument("--dest", help="import's kong admin,like http://IP:PORT")
     parser.add_argument("--host", help="migrate to new kong host")
     parser.add_argument("--project", help="Import's project, default all", default="all")
-    parser.add_argument("--profile", help="new envoirmenti, default is dev", default="dev")
+    parser.add_argument("--profile", help="new environment, default is dev", default="dev")
     parser.add_argument("--filename", help="export to filename or load from filename")
     parser.add_argument("--prefix", help="The URIS add the prefix,example /moses to /prefix/moses", default=None)
     args = parser.parse_args()
     check = Check(args)
     if check.check():
         if args.action == 'export':
+            pprint("export")
             src_kong = Kong(args.src)
             src_kong.export(args.filename)
         elif args.action == 'migrate':
             dest_kong = Kong(args.dest)
             dest_kong.migrate(args.filename, args.host, args.project, args.profile, args.prefix)
         elif args.action == 'cleanall':
-             print 'Warnning,You will delete {0} all apis'.format(args.dest)
-             print 'Warnning,You will delete {0} all apis'.format(args.dest)
-             print 'Warnning,You will delete {0} all apis'.format(args.dest)
-             time.sleep(20) 
-             # dest_kong = Kong(args.dest)
-             # dest_kong.cleanall()
+            pprint('Warning,You will delete {0} all apis'.format(args.dest))
+            pprint("Sleep 20 second")
+            time.sleep(20)
+            dest_kong = Kong(args.dest)
+            dest_kong.cleanall()
         elif args.action == "import":
-             dest_kong = Kong(args.dest)
-             dest_kong.dump(args.filename)
+            dest_kong = Kong(args.dest)
+            dest_kong.dump(args.filename)
+    else:
+        pprint("Oops, yun don't use the script right, Plase use -h to get help")
